maalicious/requirement.txt
numpy
pandas
python-whois
tld
scikit-learn
tensorflow
tf_keras
